Discover entity types while checking permissions
Entity type need to be discovered because it is unknown when using cmdb
crud rest interface. For example following request can be made:

```
DELETE https://cmdb/{entityId}
```

As a conclusion at the beginning we need to fetch entity, check its type
and next check if user has permission to perform concrete action on this
entity.